backend/requirements.txt
Django==3.2
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
Pillow==9.2.0
psycopg2-binary==2.9.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1 
requests==2.26.0
